requirements.txt
streamlit
psycopg2-binary
fuzzywuzzy
nltk
python-dotenv
python-Levenshtein

